Fixed LevelPresenter.cs Slider to have a float value

+++ Assets/DesignPatterns/MVP/LevelPresenter.cs
@@ -71,7 +71,7 @@ namespace DesignPatterns.MVP
 
         private void OnExperienceChanged()
         {
-            m_xpSlider.value = m_level.ExperiencePoints / m_level.MaxExperiencePoints;
+            m_xpSlider.value = (float)m_level.ExperiencePoints / m_level.MaxExperiencePoints;
         }
 
         private void OnLevelUp(int currentlevel)
